Chart: Use standard ordered-pair notation for scatter tooltip.
For the point X = 42, Y = 1024, the original format would
display "42 / 1024" as the tooltip text. This is a highly
non-standard format, and is thus confusing to the user.
The standard notation format for an ordered pair of
coordinates in a Cartesian coordinate system would be
"(42, 1024)". Note that the values are separated by a
comma, not by a slash. Wrapping the coordinate pair in
parentheses is not strictly necessary, but it is part
of the standard notation, and it should hopefully make
it more clear to the user that what they are looking at
in the tooltip is a Cartesian coordinate pair.