Tooltip/Chart: Ensure tracking tooltips don't go off-screen.
The TTF_ABSOLUTE flag is removed when installing a tracking
tooltip, thus allowing automatic adjustments to be made to
the display coordinates specified for the tooltip.

The specified display coordinates are still respected;
the tooltip is still displayed at approximately the
requested location on-screen, unless that would cause
the tooltip to display partially or fully off-screen,
in which case, the specified coordinates will be
automatically adjusted to ensure that the entire
tooltip fits on the screen.

Tracking tooltips are (currently) used only by the
Chart control, so this change causes no unintended
regressions within the library. And it has no
regressions with respect to the Chart control:
tooltips still behave the same way, except that
they now cannot ever extend off the screen.