Tooltip: Properly handle ampersand characters.
This change adds the TTS_NOPREFIX style flag when creating
a tooltip control, which prevents ampersand characters
from automatically being stripped from the tooltip's
display string (and also automatically terminating the
string at a tab character). This style flag should *not*
be specified when the same string is being used as both
as the display text for a tooltip and the caption for
a control/menu item. However, that is not the case here,
since the tooltip control is only used by the Chart,
Mditab, and Treelist controls to display tooltips on
items that they contain, and those items will not
contain keyboard mnemonic prefixes. Prior to this
change, if those items contained ampersands, the
text would not be displayed properly in the tooltip.