Chart: Add MC_CHN_HOTTRACK notification and update demo.
The implementation is loosely modeled on LVN_HOTTRACK
(and other similar notifications sent by Win32 controls).
The intention is to allow the parent window to display
the same information as is (or would be) displayed in
the tooltip somewhere else, like a status bar.

Since the information used by the tooltip needs to be
gathered anyway to send along with the notification,
this changeset also centralizes the tooltip display
code.