Chart: Implement interactive legend; improve color-related painting.
This changeset marks the initial implementation of an interactive legend,
where the user can click on an item (data set) in order to toggle its
state. The default state is, of course, to display the corresponding
values in the plot normally. But, when a data set is clicked, the
corresponding values in the plot are displayed grayed-out with
alpha-channel transparency (so that data sets plotted later in
the Z order do not cover up the focused data sets, even when the
former have been grayed).

This is a commonly-implemented feature of charting libraries, and
it is handy for improving the user experience with large numbers
of data sets. mCtrl's Chart control already had the very useful
feature of highlighting the corresponding values in the plot
(with an "aura", as the code calls it) when a data set item
is hovered over in the legend. This satisfies one use-case for
the interactive legend, making it possible for the user to
easily identify one data set out of many, but it does not
make it any easier to compare a subset of data sets without
interference from myriad others.

The state is tracked simply by adding a "grayed" member to the
chart_data_t structure, which contains state information for
a data series. Aside from the message-handling code, the other
changes are essentially all to the painting code, implementing
the "grayed" state.

As part of these changes, the opportunity was taken to make
two other improvements:
 (A) Factor out hard-coded color values into constants,
     making it easier to change them in the future and
     even potentially make them client-configurable.
 (B) When painting labels inside pie slices, ensure that
     the text color contrasts with the background
     (data set) color. A very basic, but workable,
     implementation is used (and also added to the
     color utility functions).