Chart: More painting improvements, including customizable fore and back colors.
Chart colors (i.e., non-data-set colors) are no longer hard-coded;
instead, they are automatically derived from background and
foreground colors using varying levels of alpha-channel opacity.
The resulting colors are identical to the original colors when
the default color scheme is used (COLOR_WINDOW for the background
color; COLOR_WINDOWTEXT for the foreground color), but because of
this new way that colors are derived, it becomes possible for the
Chart control to support user-definable colors (in addition to
having better support for users who customize their system colors
to something other than white and black, respectively).

To support user-definable colors, two new messages have been added:
MC_CHM_GETCOLORS and MC_CHM_SETCOLORS. The foreground and background
colors are set together, as a unit, rather than individually, for
several reasons: to reduce the number of messages that are required,
to improve efficiency, and because system colors (which is what a
programmer who respects user preferences will set these to) should
never be mixed. The default colors (represented by MC_CLR_DEFAULT)
are the same as the original colors: COLOR_WINDOW for the background
color and COLOR_WINDOWTEXT for the foreground color. By using the
new MC_CHM_SETCOLORS message, a client can configure the Chart
control to use COLOR_BTNFACE and COLOR_BTNTEXT, respectively; or,
of course, they can set any custom color scheme that they want.
With the changes to the painting logic, all of these look good.

The example project for the Chart control has been updated to
draw the scatter plot (which already uses client-defined colors
for its data sets) using client-defined colors for the background
and foreground.