Chart: Rearrange example charts to facilitate comparisons between types.
The line, area, and stacked-area charts are similar,
so this new layout has them all in the first column.
The column and stacked-column remain together in the
second column, just as the bar and stacked-bar remain
together in the third column. The scatter and pie
charts are now inserted at the end (rather than at
the beginning); the scatter remains in the middle,
while the pie remains at the end.