Chart: Implement MC_CHS_CONNECTEDSCATTER style.
This style produces a chart that is exactly the same as a standard
scatter chart (MC_CHS_SCATTER), except that the points are all
connected by lines, exactly as with a line chart (MC_CHS_LINE).
A line chart with the same X-axis delta and step size would look
exactly the same as a connected-scatter chart.