Chart: Make line charts with a dense X-axis readable.
The X axis of line charts is the only axis with the type
GRID_AXIS_DISCRETE. Previously, axes of this type had
their grid_delta forced to 1. However, this can result
in a chart that is unreadable and unusable if there are
many values on the X axis (e.g., 0 to 1000) and/or if
the chart's width is very narrow. There is no reason
that the X axis of line charts should be treated any
differently from the X axis of scatter charts when it
comes to drawing the gridlines (and therefore labels).
All discrete values are still plotted; the only change
is the spacing of the X-axis gridlines (if they are
configured to be drawn) and, therefore, the labels.