Chart: Fix display of negative values greater than -1.
For values between 0 and -1, the negative sign was not being printed
because it was not included as part of the string. The division of
the value by the factor was positive in this case, since integers
don't support a representation for -0, so there was no leading
negative sign included in the string.